refactor: extract runtime episode runner
